Kill MutabilityFlags.IgnoreImmutabilityAttribute
This means that UnsafeStaticsAnalyzer will also check [Immutable] types
within the same assembly.

This is duplicated effort but it will be addressed by making
UnsafeStaticsAnalyzer and ImmutablityAnalyzer share a cache.

What this buys us is less flags. This is important for caching (flags
are effectively part of the cache key.)

In the next PR I'm going to kill the remaining
MutabilityFlags.AllowUnsealed in favour of an overload.